BREAKING CHANGE: options with no link now default to looping back to the options section, instead of quitting the interpreter. The old behaviour can be restored in a questmark document via an options flag.
This makes more sense for the kinds of conversational texts where you want to give users the option to "interrogate" a character, without them repeating the same introduction text over and over.

For quests, you can explicitly have an option link back to the same section it's part of, which will cause the entire section text to be displayed again.